==> notebooks_data_processing/1_make_dataset.ipynb <==
# Data Exploration and Introduction

### import packages

import pandas as pd
pd.options.mode.chained_assignment = None  # default='warn'
import pickle
import re
%matplotlib inline

### import helper functions


import os
import sys
sys.path.insert(0, os.path.abspath(os.path.join('..', 'src')))

'''helper functions'''
import make_dataset as mk # make dataset - data processing

import make_dataset_PIMS_API as mkp 

# Import API data:

# uses neural translation for french and spanish logframes. Takes approx 4h withoug GPU support.
#df_api, df_logs = mk.import_api_data(all_timelines = True)

# Process Data

## Process taxonomy data

# use helper functios from make_dataset to import tagging table:
taxonomy = mk.import_raw_data('tagging_table_v17') 
print('')

# use helper function to import and process indicators:
indicators = mk.import_indicators()
print('')

# create datasets: processed data, dataframe with only raw texts and download data from PIMS API to compare and fill in:
taxonomy_processed, pims = mk.create_training_texts(taxonomy, compare_with_API = True, under_implementation = True, replace_short = True)

print('save data')
taxonomy_processed.to_excel(os.path.abspath(os.path.join('..', 'data/interim'))+'/taxonomy_processed.xlsx', index=False)

### save list of PIMS that have no text in taxonomy:

taxonomy_no_pims, pims = mk.create_training_texts(taxonomy, compare_with_API = False, under_implementation = True)
taxonomy_filtered_no_pims, data = mk.create_subset(taxonomy_no_pims, "all_text")
print('xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx')
print(taxonomy_no_pims.shape)
print(taxonomy_filtered_no_pims.shape)

# taxonomy_no_pims['all_text'] = taxonomy_no_pims['all_text'].fillna('').astype(str)
# df1 =  taxonomy_no_pims[taxonomy_no_pims['all_text'].str.contains('[A-Za-z]')]

# total_merge = taxonomy_no_pims.merge(df1, on='PIMS_ID', how='outer', indicator=True)
# total_merge['all_text_y'] = total_merge['all_text_y'].fillna('').astype(str)

# R_1 = total_merge[total_merge['_merge']=='both']
# R_2 = total_merge[total_merge['_merge']=='left_only']
# R_3 = total_merge[total_merge['_merge']=='right_only']

##  detect missing and faulty logframes

##save flaged columns and texts:
# flag = taxonomy_processed[['PIMS_ID', 'all_text', 'description', 'pims_description', 'pims_objectives','pims_outcomes', 'pims_text','title_only']]
# flag = flag.loc[flag['title_only'] == True]
# flag = flag.reset_index(drop=True)
# flag = flag[['PIMS_ID','pims_text', 'title_only']]
# flag['pims_text'] = flag['pims_text'].fillna('not available')

# flag = flag.rename(columns={"title_only": "available_from_pims"})

# for index, row in flag.iterrows():
#     if row['pims_text'] == 'not available':
#         flag.at[index, 'available_from_pims'] = False
        
#     if len(row['pims_text']) < 600 and row['available_from_pims'] == True:
#         flag.at[index, 'available_but_incomplete'] = True
#     else:
#         flag.at[index, 'available_but_incomplete'] = False
           
# flag.to_excel(os.path.abspath(os.path.join('..', 'data/temp'))+'/missing_logframes.xlsx')    

# # not available or if available non useable list:
# flag_2 = flag
# for index, row in flag_2.iterrows():
    
#     if row['available_from_pims'] == True:
#         if row['available_but_incomplete'] == False:
#             flag_2.drop(index, inplace=True)
# flag_2 = flag_2.reset_index(drop=True) 
           
# flag_2.to_excel(os.path.abspath(os.path.join('..', 'data/temp'))+'/checkup.xlsx') 

# #no logs
# no_logs  = flag[flag.available_from_pims == False]
# no_logs = no_logs.reset_index(drop=True)

# no_logs = no_logs[['PIMS_ID', 'pims_text', 'available_from_pims']]

# no_logs.to_excel(os.path.abspath(os.path.join('..', 'data/temp'))+'/no_logframes.xlsx')

# #logs but seems incomplete:
# incomplete  = flag[flag.available_but_incomplete == True]
# incomplete = incomplete.reset_index(drop=True)

# incomplete = incomplete[['PIMS_ID', 'pims_text', 'available_but_incomplete']]

# incomplete.to_excel(os.path.abspath(os.path.join('..', 'data/temp'))+'/incomplete.xlsx')

### Focus on most commonly appearing text: 'all_text'

taxonomy_filtered, data = mk.create_subset(taxonomy_processed, "all_text")

print('save data...')
taxonomy_filtered.to_excel(os.path.abspath(os.path.join('..', 'data/interim'))+'/taxonomy_filtered.xlsx', index=False)

### check lenght of text columns for training data:

import plotly.express as px
all_text, data = mk.create_subset(taxonomy_filtered, "all_text")
all_text['all_text_len'] = all_text['all_text'].apply(len)
print('')
print('')
logframe, data = mk.create_subset(taxonomy_filtered, "logframe")
logframe['logframe_len'] = logframe['logframe'].apply(len) 
print('')
print('')
description, data = mk.create_subset(taxonomy_filtered, "description")
description['description_len'] = description['description'].apply(len)
print('')
print('')
objectives, data = mk.create_subset(taxonomy_filtered, "objectives")
objectives['objectives_len'] = objectives['objectives'].apply(len)
print('')
print('')
outcomes, data = mk.create_subset(taxonomy_filtered, "outcomes")
outcomes['outcomes_len'] = objectives['outcomes'].apply(len)
print('')
print('')
outputs, data = mk.create_subset(taxonomy_filtered, "outputs")
outputs['outcomes_len'] = outputs['outputs'].apply(len)
print('')
print('')
indicator, data = mk.create_subset(taxonomy_filtered, "Indicator")
indicator['indicator_len'] = outputs['Indicator'].apply(len)
print('')
print('')
pims_text, data = mk.create_subset(taxonomy_filtered, "pims_text")
pims_text['pims_text_len'] = pims_text['pims_text'].apply(len)

### description:
all_text.hist(column='all_text_len', bins=100)

from matplotlib_venn import venn2
import matplotlib.pyplot as plt
fig = plt.figure()
a = all_text.PIMS_ID.to_list()
b = logframe.PIMS_ID.to_list()
c = description.PIMS_ID.to_list()
d = objectives.PIMS_ID.to_list()
e = outputs.PIMS_ID.to_list()
f = outcomes.PIMS_ID.to_list()

set1 = set(a)
set2 = set(b)
set3 = set(c)
set4 = set(d)
set5 = set(e)
set6 = set(f)

v = venn2([set1, set6], ('all_text', 'outcomes'))

taxonomy_filtered.columns


labels_only = pd.DataFrame(taxonomy_filtered, columns=[ 'technical_team', 'cross-cutting_programme_1', 'technical_area_1',
       'technical_sub-area_1', 'cross-cutting_programme_2', 'technical_area_2',
       'technical_sub-area_2', 'transformed_sectors',
       'landscapes_1', 'sub-landscapes_1', 'landscapes_2', 'sub-landscapes_2',
       'landscapes_3', 'sub-landscapes_3', 'sdg_goal_reference_from_pims+',
       'sdg_goal', 'sdg_targets_(top_three)', 'sdg_goal_summary', 'undp_roles',
       'main_theme_', 'sub-_strategy_1', 'main_strategy_2', 'sub-_strategy_2',
       'main_strategy_3', 'sub-_strategy_3', 'pathways', 'targeted_risks',
       '*conventions/protocols/plans', '*social_inclusion_&_engagement',
       '*gender_equality', '*types_of_private_sector', '*hot_topics'])

multi_class = pd.DataFrame(taxonomy_filtered, columns=['technical_team'])

multi_label_upper_3 = pd.DataFrame(taxonomy_filtered, columns=['undp_roles', 'pathways',
       'targeted_risks', '*conventions/protocols/plans',
       '*social_inclusion_&_engagement', '*gender_equality',
       '*types_of_private_sector', '*hot_topics'])


sequential_multi_class_multi_label = pd.DataFrame(taxonomy_filtered, columns=['main_strategy_1',  'main_strategy_2', 'main_strategy_3',  'landscapes_1','landscapes_2', 'landscapes_3']) 

multi_label_hierarchical_upper_2 = pd.DataFrame(taxonomy_filtered, columns=['technical_sub-area_1', 'technical_sub-area_2']) 

multi_label_hierarchical_upper_3 = pd.DataFrame(taxonomy_filtered, columns=['technical_sub-sub_area_1', 'technical_sub-sub_area_2', 
                                                                              'sub-landscapes', 'sub-landscapes_2', 'sub-landscapes_3',  
                                                                             'sub-_strategy_1', 'sub-_strategy_2', 'sub-_strategy_3']) 

multi_label_few_hierarchical = pd.DataFrame(taxonomy_filtered, columns=['transformed_sectors']) 

for index, row in taxonomy_filtered.iterrows():
    
    if row['technical_team'] == 'Climate_Change_Mitigation_Programme':
        print(row['PIMS_ID'])

## Multi-Class Categorisation: Categorical charts

for cate in multi_label_upper_3:
    class_ = vis.disp_category(cate, taxonomy_filtered, min_items=1)

## Sequential multi label/multi class: Categorical charts


for cate in sequential_multi_class_multi_label:
    print (cate)
    class_ = vis.plotly_pie(cate, taxonomy_filtered, min_items=1)

## Radar chart

for cate in multi_class:
    print (cate)
    class_ = vis.plotly_radar(cate, taxonomy_filtered, min_items=1)

# Summary:

* Very small training data set.
* Highly skewed distribution of training data set. 
* Highly unbalanced classification problems. 
* Large number of labels to predict.
* Hierarchical structures.
* Additional rules - "only up to two/three categories".
