==> notebooks_data_processing/2_data_cleaning.ipynb <==
import pandas as pd
import pickle

import os
import sys
sys.path.insert(0, os.path.abspath(os.path.join('..', 'src')))

#helper functions
import clean_dataset as clean

import make_dataset as mk

taxonomy_filtered = pd.read_excel(os.path.abspath(os.path.join('..', 'data/interim'))+'/taxonomy_filtered.xlsx')    

taxonomy_filtered.prodoc_link

'''translate non-english logframes'''

'''translate remaining french entries'''
taxonomy_filtered = clean.detect_lang(taxonomy_filtered, "all_text", "fr", 
                                      printing =True, translate = True, remove = False)

'''translate remaining spanish entries'''
taxonomy_filtered = clean.detect_lang(taxonomy_filtered, "all_text", "es", 
                                      printing =True, translate = True, remove = False)

"""clean text"""

columns_to_clean = ['all_text']

for column in columns_to_clean:
    taxonomy_filtered[column] = taxonomy_filtered[column].astype(str)
    
    #basic cleaning
    new = column + "_clean"
    print(new)
    taxonomy_filtered[new] = taxonomy_filtered[column].apply(clean.basic)
    
    #lemmatise and stemming + basic cleaning
    new_spacy = column + "_clean_spacy"
    print(new_spacy)
    taxonomy_filtered[new_spacy] = taxonomy_filtered[column].apply(clean.spacy_clean)

"""detect and remove short texts"""

taxonomy_filtered = mk.truncate(taxonomy_filtered, 300, save=False, printing =True)

'''clean columns a little:'''
taxonomy_filtered.columns = taxonomy_filtered.columns.str.replace("-", "_")
taxonomy_filtered.columns = taxonomy_filtered.columns.str.replace("__", "_")
taxonomy_filtered.columns = taxonomy_filtered.columns.str.replace("*", "_")

taxonomy_filtered.to_excel(os.path.abspath(os.path.join('..', 'data/processed'))+'/taxonomy_clean.xlsx', index=False)
