==> notebooks_data_processing/3_make_dataset_encoding_labels.ipynb <==
import os
import sys
sys.path.insert(0, os.path.abspath(os.path.join('..', 'src')))
import pandas as pd
import spacy
import make_dataset as mk
import pickle

'''import data'''
taxonomy = pd.read_excel(os.path.abspath(os.path.join('..', 'data/processed'))+'/taxonomy_clean.xlsx')

taxonomy.columns

## Clean up labels for the highest priorty categories:

'''Clean labels for social inclusion & engagement'''

taxonomy['_social_inclusion_&_engagement'] = taxonomy['_social_inclusion_&_engagement'].str.replace('Indigenous peoples', 'Indigenous')
taxonomy['_social_inclusion_&_engagement'] = taxonomy['_social_inclusion_&_engagement'].str.replace('Indigenous', 'Indigenous peoples')
taxonomy['_social_inclusion_&_engagement'] = taxonomy['_social_inclusion_&_engagement'].str.replace('Waste pickers', 'Waste Picker')

taxonomy['_social_inclusion_&_engagement'] = taxonomy['_social_inclusion_&_engagement'].str.replace('\nWomen\n', 'Women')


'''Clean labels for hot topics'''
# one to clean - only Unnamed: 4 remains a mystery.
taxonomy['_hot_topics'] = taxonomy['_hot_topics'].str.replace('Reduce poverty', 'Poverty Reduction')
                                  

df3, df2, tt = mk.labeling_encoding(taxonomy, [ 'technical_team', 'cross_cutting_programme_1', 'technical_area_1',
       'technical_sub_area_1', 'cross_cutting_programme_2', 'technical_area_2',
       'technical_sub_area_2', 'transformed_sectors',
       'landscapes_1', 'sub_landscapes_1', 'landscapes_2', 'sub_landscapes_2',
       'landscapes_3', 'sub_landscapes_3', 'undp_roles',
       'main_strategy_1', 'sub_strategy_1', 'main_strategy_2', 'sub_strategy_2',
       'main_strategy_3', 'sub_strategy_3', 'pathways', 'targeted_risks',
       '_conventions/protocols/plans', '_social_inclusion_&_engagement',
       '_gender_equality', '_types_of_private_sector', '_hot_topics'], save = True)

## Merged strategies

'''import data'''
strategy = pd.read_excel(os.path.abspath(os.path.join('..', 'data/raw'))+'/merged_tagging_table_v1.xlsx')
strategy = strategy[['PIMS #', 'MAIN STRATEGY 1', 'First.8','Second.8','Third.8','Rest.7','MAIN STRATEGY 2','First.9',
                     'Second.9','Third.9','Rest.8','MAIN STRATEGY 3','First.10','Second.10','Third.10','Rest.9']]

strategy.columns = strategy.columns.str.replace(' ', '_').str.lower()
strategy = strategy.rename(columns={"pims_#": "PIMS_ID"})
strategy = strategy[strategy['PIMS_ID'].notna()]
strategy = strategy[strategy['main_strategy_1'].notna()]
strategy = strategy.fillna('')

df3, df2, tt = mk.labeling_encoding(strategy, ['main_strategy_1', 'first.8', 'second.8', 'third.8',
       'rest.7', 'main_strategy_2', 'first.9', 'second.9', 'third.9', 'rest.8',
       'main_strategy_3', 'first.10', 'second.10', 'third.10', 'rest.9'], save = True)
